重构：平台管理-将登录页背景与玻璃卡片的设置，移动到“主题与样式管理”
