+++ ui/src/views/admin/admin-settings-base.vue
@@ -3,17 +3,12 @@ import { useChatStore } from '@/stores/chat';
 import { useUtilsStore } from '@/stores/utils';
 import type { ServerConfig, BackupInfo } from '@/types';
 import { Message } from '@vicons/tabler';
-import { Photo as ImageIcon, X } from '@vicons/tabler';
 import { cloneDeep } from 'lodash-es';
-import { NIcon, useMessage, NButton, NPopconfirm } from 'naive-ui';
+import { useMessage, NButton, NPopconfirm } from 'naive-ui';
 import { computed, nextTick } from 'vue';
 import { onMounted, ref, watch, h } from 'vue';
 import { api } from '@/stores/_config';
 import dayjs from 'dayjs';
-import { resolveAttachmentUrl } from '@/composables/useAttachmentResolver';
-import { uploadImageAttachment } from '@/views/chat/composables/useAttachmentUploader';
-import { useImageCompressor } from '@/composables/useImageCompressor';
-import { useLoginGlass } from '@/composables/useLoginGlass';
 
 const chat = useChatStore();
 
@@ -645,260 +640,6 @@ const sendSmtpTestEmail = async () => {
   }
 }
 
-// Login background state
-const { compress: compressImage } = useImageCompressor();
-const loginBgUploading = ref(false);
-const loginBgFileInput = ref<HTMLInputElement | null>(null);
-
-const loginBgAttachmentId = computed({
-  get: () => model.value.loginBackground?.attachmentId || '',
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.attachmentId = val;
-  },
-});
-
-const loginBgMode = computed({
-  get: () => model.value.loginBackground?.mode || 'cover',
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.mode = val;
-  },
-});
-
-const loginBgOpacity = computed({
-  get: () => model.value.loginBackground?.opacity ?? 30,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.opacity = val;
-  },
-});
-
-const loginBgBlur = computed({
-  get: () => model.value.loginBackground?.blur ?? 0,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.blur = val;
-  },
-});
-
-const loginBgBrightness = computed({
-  get: () => model.value.loginBackground?.brightness ?? 100,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.brightness = val;
-  },
-});
-
-const loginBgOverlayColor = computed({
-  get: () => model.value.loginBackground?.overlayColor || '',
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.overlayColor = val;
-  },
-});
-
-const loginBgOverlayOpacity = computed({
-  get: () => model.value.loginBackground?.overlayOpacity ?? 0,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.overlayOpacity = val;
-  },
-});
-
-const loginPanelAutoTint = computed({
-  get: () => model.value.loginBackground?.panelAutoTint ?? true,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.panelAutoTint = val;
-  },
-});
-
-const loginPanelTintColor = computed({
-  get: () => model.value.loginBackground?.panelTintColor || '',
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.panelTintColor = val;
-  },
-});
-
-const loginPanelTintOpacity = computed({
-  get: () => model.value.loginBackground?.panelTintOpacity ?? 72,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.panelTintOpacity = val;
-  },
-});
-
-const loginPanelBlur = computed({
-  get: () => model.value.loginBackground?.panelBlur ?? 14,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.panelBlur = val;
-  },
-});
-
-const loginPanelSaturate = computed({
-  get: () => model.value.loginBackground?.panelSaturate ?? 120,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.panelSaturate = val;
-  },
-});
-
-const loginPanelContrast = computed({
-  get: () => model.value.loginBackground?.panelContrast ?? 105,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.panelContrast = val;
-  },
-});
-
-const loginPanelBorderOpacity = computed({
-  get: () => model.value.loginBackground?.panelBorderOpacity ?? 18,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.panelBorderOpacity = val;
-  },
-});
-
-const loginPanelShadowStrength = computed({
-  get: () => model.value.loginBackground?.panelShadowStrength ?? 22,
-  set: (val) => {
-    if (!model.value.loginBackground) {
-      model.value.loginBackground = {};
-    }
-    model.value.loginBackground.panelShadowStrength = val;
-  },
-});
-
-const loginBgUrl = computed(() => {
-  const id = loginBgAttachmentId.value;
-  if (!id) return '';
-  return resolveAttachmentUrl(id.startsWith('id:') ? id : `id:${id}`);
-});
-
-const loginBgModeOptions = [
-  { label: '铺满 (Cover)', value: 'cover' },
-  { label: '适应 (Contain)', value: 'contain' },
-  { label: '平铺 (Tile)', value: 'tile' },
-  { label: '居中 (Center)', value: 'center' },
-];
-
-const loginBgPreviewStyle = computed(() => {
-  if (!loginBgUrl.value) return {};
-  const mode = loginBgMode.value;
-  let bgSize = 'cover';
-  let bgRepeat = 'no-repeat';
-  let bgPosition = 'center';
-  switch (mode) {
-    case 'contain':
-      bgSize = 'contain';
-      break;
-    case 'tile':
-      bgSize = 'auto';
-      bgRepeat = 'repeat';
-      break;
-    case 'center':
-      bgSize = 'auto';
-      bgPosition = 'center';
-      break;
-  }
-  return {
-    backgroundImage: `url(${loginBgUrl.value})`,
-    backgroundSize: bgSize,
-    backgroundRepeat: bgRepeat,
-    backgroundPosition: bgPosition,
-    opacity: loginBgOpacity.value / 100,
-    filter: `blur(${loginBgBlur.value}px) brightness(${loginBgBrightness.value}%)`,
-  };
-});
-
-const loginBgOverlayStyle = computed(() => {
-  if (!loginBgOverlayColor.value || !loginBgOverlayOpacity.value) return null;
-  return {
-    backgroundColor: loginBgOverlayColor.value,
-    opacity: loginBgOverlayOpacity.value / 100,
-  };
-});
-
-const { glassStyle: loginGlassStyle } = useLoginGlass({
-  imageUrl: loginBgUrl,
-  config: computed(() => model.value.loginBackground),
-  enabled: computed(() => !!loginBgUrl.value),
-  radius: '8px',
-});
-
-const loginGlassPreviewStyle = computed(() => ({
-  ...loginGlassStyle.value,
-  '--sc-glass-radius': '8px',
-}));
-
-const triggerLoginBgUpload = () => {
-  loginBgFileInput.value?.click();
-};
-
-const handleLoginBgFileChange = async (event: Event) => {
-  const input = event.target as HTMLInputElement;
-  const file = input?.files?.[0];
-  if (!file) return;
-  input.value = '';
-
-  const sizeLimit = utils.fileSizeLimit;
-  if (file.size > sizeLimit) {
-    const limitMB = (sizeLimit / 1024 / 1024).toFixed(1);
-    message.error(`文件大小超过限制（最大 ${limitMB} MB）`);
-    return;
-  }
-
-  loginBgUploading.value = true;
-  try {
-    const compressed = await compressImage(file, { maxWidth: 1920, maxHeight: 1080 });
-    const result = await uploadImageAttachment(compressed, { channelId: 'login-background', skipCompression: true });
-    let attachId = result.attachmentId || '';
-    if (attachId.startsWith('id:')) {
-      attachId = attachId.slice(3);
-    }
-    loginBgAttachmentId.value = attachId;
-    message.success('背景图片上传成功');
-  } catch (err: any) {
-    message.error(err?.message || '上传失败');
-  } finally {
-    loginBgUploading.value = false;
-  }
-};
-
-const clearLoginBg = () => {
-  loginBgAttachmentId.value = '';
-};
 </script>
 
 <template>
@@ -985,121 +726,7 @@ const clearLoginBg = () => {
         <n-input-number v-model:value="model.keywordMaxLength" :min="100" :max="10000" />
       </n-form-item>
 
-      <!-- Login Background Section -->
       <n-collapse class="settings-collapse" :default-expanded-names="[]">
-        <n-collapse-item title="登录页背景" name="login-bg">
-          <input ref="loginBgFileInput" type="file" accept="image/*" class="hidden" @change="handleLoginBgFileChange">
-          <n-form-item label="背景图片">
-            <div class="flex gap-3 items-center">
-              <div
-                class="login-bg-no-option"
-                :class="{ active: !loginBgAttachmentId }"
-                @click="clearLoginBg"
-              >
-                <NIcon :component="X" :size="16" />
-                <span>无</span>
-              </div>
-              <div v-if="loginBgUrl" class="login-bg-thumb-wrapper">
-                <img :src="loginBgUrl" alt="登录背景" class="login-bg-thumb" />
-              </div>
-              <n-button size="small" :loading="loginBgUploading" @click="triggerLoginBgUpload">
-                <template #icon><NIcon :component="ImageIcon" /></template>
-                {{ loginBgUrl ? '更换' : '上传' }}
-              </n-button>
-            </div>
-          </n-form-item>
-
-          <template v-if="loginBgAttachmentId">
-            <n-form-item label="显示模式">
-              <n-select v-model:value="loginBgMode" :options="loginBgModeOptions" style="width: 180px;" />
-            </n-form-item>
-            <n-form-item label="透明度">
-              <div class="login-bg-control-row">
-                <n-slider class="login-bg-slider" v-model:value="loginBgOpacity" :min="0" :max="100" :step="1" :tooltip="false" />
-                <span class="login-bg-value">{{ loginBgOpacity }}%</span>
-              </div>
-            </n-form-item>
-            <n-form-item label="模糊度">
-              <div class="login-bg-control-row">
-                <n-slider class="login-bg-slider" v-model:value="loginBgBlur" :min="0" :max="20" :step="1" :tooltip="false" />
-                <span class="login-bg-value">{{ loginBgBlur }}px</span>
-              </div>
-            </n-form-item>
-            <n-form-item label="亮度">
-              <div class="login-bg-control-row">
-                <n-slider class="login-bg-slider" v-model:value="loginBgBrightness" :min="50" :max="150" :step="1" :tooltip="false" />
-                <span class="login-bg-value">{{ loginBgBrightness }}%</span>
-              </div>
-            </n-form-item>
-            <n-form-item label="叠加层颜色">
-              <n-color-picker v-model:value="loginBgOverlayColor" :show-alpha="false" style="width: 100px;" />
-              <n-button v-if="loginBgOverlayColor" size="tiny" quaternary class="ml-2" @click="loginBgOverlayColor = ''">清除</n-button>
-            </n-form-item>
-            <n-form-item v-if="loginBgOverlayColor" label="叠加层透明度">
-              <div class="login-bg-control-row">
-                <n-slider class="login-bg-slider" v-model:value="loginBgOverlayOpacity" :min="0" :max="100" :step="1" :tooltip="false" />
-                <span class="login-bg-value">{{ loginBgOverlayOpacity }}%</span>
-              </div>
-            </n-form-item>
-          </template>
-        </n-collapse-item>
-
-        <n-collapse-item v-if="loginBgAttachmentId" title="玻璃卡片" name="login-glass">
-          <n-form-item label="自动色调">
-            <n-switch v-model:value="loginPanelAutoTint" />
-          </n-form-item>
-          <n-form-item label="玻璃色">
-            <n-color-picker v-model:value="loginPanelTintColor" :show-alpha="false" style="width: 100px;" />
-            <n-button v-if="loginPanelTintColor" size="tiny" quaternary class="ml-2" @click="loginPanelTintColor = ''">清除</n-button>
-          </n-form-item>
-          <n-form-item label="玻璃透明度">
-            <div class="login-bg-control-row">
-              <n-slider class="login-bg-slider" v-model:value="loginPanelTintOpacity" :min="30" :max="95" :step="1" :tooltip="false" />
-              <span class="login-bg-value">{{ loginPanelTintOpacity }}%</span>
-            </div>
-          </n-form-item>
-          <n-form-item label="玻璃模糊">
-            <div class="login-bg-control-row">
-              <n-slider class="login-bg-slider" v-model:value="loginPanelBlur" :min="0" :max="30" :step="1" :tooltip="false" />
-              <span class="login-bg-value">{{ loginPanelBlur }}px</span>
-            </div>
-          </n-form-item>
-          <n-form-item label="饱和度">
-            <div class="login-bg-control-row">
-              <n-slider class="login-bg-slider" v-model:value="loginPanelSaturate" :min="80" :max="180" :step="1" :tooltip="false" />
-              <span class="login-bg-value">{{ loginPanelSaturate }}%</span>
-            </div>
-          </n-form-item>
-          <n-form-item label="对比度">
-            <div class="login-bg-control-row">
-              <n-slider class="login-bg-slider" v-model:value="loginPanelContrast" :min="90" :max="140" :step="1" :tooltip="false" />
-              <span class="login-bg-value">{{ loginPanelContrast }}%</span>
-            </div>
-          </n-form-item>
-          <n-form-item label="边框强度">
-            <div class="login-bg-control-row">
-              <n-slider class="login-bg-slider" v-model:value="loginPanelBorderOpacity" :min="0" :max="60" :step="1" :tooltip="false" />
-              <span class="login-bg-value">{{ loginPanelBorderOpacity }}%</span>
-            </div>
-          </n-form-item>
-          <n-form-item label="阴影强度">
-            <div class="login-bg-control-row">
-              <n-slider class="login-bg-slider" v-model:value="loginPanelShadowStrength" :min="0" :max="60" :step="1" :tooltip="false" />
-              <span class="login-bg-value">{{ loginPanelShadowStrength }}%</span>
-            </div>
-          </n-form-item>
-          <n-form-item label="预览">
-            <div class="login-bg-preview">
-              <div class="login-bg-preview-layer" :style="loginBgPreviewStyle"></div>
-              <div v-if="loginBgOverlayStyle" class="login-bg-preview-overlay" :style="loginBgOverlayStyle"></div>
-              <div class="login-bg-preview-card sc-glass-panel" :style="loginGlassPreviewStyle">
-                <div class="login-bg-preview-input"></div>
-                <div class="login-bg-preview-input"></div>
-                <div class="login-bg-preview-btn"></div>
-              </div>
-            </div>
-          </n-form-item>
-        </n-collapse-item>
         <n-collapse-item title="SQLite空间压缩" name="sqlite-space-compress">
             <n-form-item label="当前数据库大小">
               <div class="flex flex-col gap-1">
@@ -1332,136 +959,5 @@ const clearLoginBg = () => {
 
 .settings-collapse {
   width: 100%;
-}
-
-/* Login Background Styles */
-.login-bg-no-option {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 48px;
-  height: 48px;
-  border: 2px dashed #d1d5db;
-  border-radius: 6px;
-  cursor: pointer;
-  transition: all 0.2s;
-  font-size: 12px;
-  color: #9ca3af;
-}
-.login-bg-no-option:hover {
-  border-color: #9ca3af;
-  color: #6b7280;
-}
-.login-bg-no-option.active {
-  border-color: #3b82f6;
-  background-color: #eff6ff;
-  color: #3b82f6;
-}
-.dark .login-bg-no-option {
-  border-color: #4b5563;
-  color: #6b7280;
-}
-.dark .login-bg-no-option:hover {
-  border-color: #6b7280;
-  color: #9ca3af;
-}
-.dark .login-bg-no-option.active {
-  border-color: #3b82f6;
-  background-color: #1e3a5f;
-  color: #60a5fa;
-}
-
-.login-bg-thumb-wrapper {
-  width: 48px;
-  height: 48px;
-  border-radius: 6px;
-  overflow: hidden;
-  border: 2px solid #3b82f6;
-}
-.login-bg-thumb {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.login-bg-preview {
-  position: relative;
-  width: 240px;
-  height: 160px;
-  border-radius: 8px;
-  overflow: hidden;
-  background-color: #f3f4f6;
-}
-.dark .login-bg-preview {
-  background-color: #1f2937;
-}
-
-.login-bg-preview-layer {
-  position: absolute;
-  inset: 0;
-  z-index: 0;
-}
-
-.login-bg-preview-overlay {
-  position: absolute;
-  inset: 0;
-  z-index: 1;
-}
-
-.login-bg-preview-card {
-  position: absolute;
-  inset: 18px 24px;
-  z-index: 2;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  padding: 12px;
-}
-
-.login-bg-preview-input {
-  width: 100%;
-  height: 18px;
-  background: rgba(255, 255, 255, 0.65);
-  border-radius: 4px;
-}
-.dark .login-bg-preview-input {
-  background: rgba(15, 23, 42, 0.6);
-}
-
-.login-bg-preview-btn {
-  width: 60%;
-  height: 20px;
-  background: rgba(59, 130, 246, 0.9);
-  border-radius: 4px;
-  margin-top: 4px;
-}
-
-.login-bg-control-row {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  min-width: 0;
-  width: 100%;
-}
-
-.login-bg-slider {
-  flex: 1;
-  min-width: 200px;
-}
-
-.login-bg-value {
-  display: inline-block;
-  width: 56px;
-  margin-left: 8px;
-  font-size: 12px;
-  color: #6b7280;
-  text-align: right;
-  font-variant-numeric: tabular-nums;
-}
-.dark .login-bg-value {
-  color: #9ca3af;
 }
 </style>
